fix(pr): add jest & fix pr comments
