+++ package.json
@@ -41,6 +41,7 @@
     "@storybook/addon-ondevice-controls": "^6.5.6",
     "@storybook/react-native": "^6.5.6",
     "@tsconfig/react-native": "^3.0.0",
+    "@types/jest": "^29.5.7",
     "@types/react": "^18.0.24",
     "@types/react-test-renderer": "^18.0.0",
     "@types/styled-components": "^5.1.27",
