+++ src/features/phone-input/lib/phone-formatter/index.ts
@@ -1,2 +1,1 @@
-export { validatePhone } from "./validate-phone"
 export { formatPhonenumber } from "./phone-formatter"
